requirements.txt
google-generativeai>=0.3.0
tenacity>=8.0.0
python-dotenv>=0.19.0
pydantic>=2.0.0 